:triangular_flag_on_post: 计算表达式求值

+++ DataSrtucture/chapter2/expression_evaluation.h
@@ -5,6 +5,6 @@
 #ifndef MAIN_EXPRESSION_EVALUATION_H
 #define MAIN_EXPRESSION_EVALUATION_H
 
-float evaluate_expression(char str[]);
+double evaluate_expression(char str[]);
 
 #endif //MAIN_EXPRESSION_EVALUATION_H
